feat: 예약 생성 api 구현 (공휴일 api 호출 실패)

+++ src/main/java/finalmission/ballparkreservation/reservation/dto/ReservationCreateResponse.java
@@ -1,6 +1,6 @@
 package finalmission.ballparkreservation.reservation.dto;
 
 public record ReservationCreateResponse(
-        Long amount
+        int amount
 ) {
 }
